move layers state related files to global directory
