+++ src/modules/baseLayer/components/switcher/BaseLayerSwitcher.js
@@ -1,7 +1,7 @@
 import { html, nothing } from 'lit-html';
 import { $injector } from '../../../../injection';
 import { BaElement } from '../../../BaElement';
-import { addLayer, removeLayer } from '../../../map/store/layers.action';
+import { addLayer, removeLayer } from '../../../../store/layers/layers.action';
 import css from './baseLayerSwitcher.css';
 
 /**
