+++ src/modules/map/store/GeolocationPlugin.js
@@ -2,7 +2,7 @@ import { observe } from '../../../utils/storeUtils';
 import { $injector } from '../../../injection';
 import { setPosition, setAccuracy, setDenied, setTracking } from './geolocation.action';
 import { changeCenter, setFit } from '../../../store/position/position.action';
-import { addLayer, removeLayer } from './layers.action';
+import { addLayer, removeLayer } from '../../../store/layers/layers.action';
 import { BaPlugin } from '../../../store/BaPlugin';
 
 /**
