+++ test/modules/footer/components/baseLayerInfo/BaseLayerInfo.test.js
@@ -1,8 +1,8 @@
 import { BaseLayerInfo } from '../../../../../src/modules/footer/components/baseLayerInfo/BaseLayerInfo';
 import { TestUtils } from '../../../../test-utils.js';
-import { layersReducer, defaultLayerProperties } from '../../../../../src/modules/map/store/layers.reducer';
+import { layersReducer, defaultLayerProperties } from '../../../../../src/store/layers/layers.reducer';
 import { positionReducer } from '../../../../../src/store/position/position.reducer';
-import { addLayer, removeLayer, modifyLayer } from '../../../../../src/modules/map/store/layers.action';
+import { addLayer, removeLayer, modifyLayer } from '../../../../../src/store/layers/layers.action';
 import { WMTSGeoResource } from '../../../../../src/services/domain/geoResources';
 import { $injector } from '../../../../../src/injection';
