+++ test/modules/map/components/attributionInfo/AttributionInfo.test.js
@@ -1,8 +1,8 @@
 import { AttributionInfo } from '../../../../../src/modules/map/components/attributionInfo/AttributionInfo';
 import { TestUtils } from '../../../../test-utils.js';
-import { layersReducer, defaultLayerProperties } from '../../../../../src/modules/map/store/layers.reducer';
+import { layersReducer, defaultLayerProperties } from '../../../../../src/store/layers/layers.reducer';
 import { positionReducer } from '../../../../../src/store/position/position.reducer'; 
-import { addLayer, removeLayer, modifyLayer } from '../../../../../src/modules/map/store/layers.action';
+import { addLayer, removeLayer, modifyLayer } from '../../../../../src/store/layers/layers.action';
 import { changeZoom } from '../../../../../src/store/position/position.action';
 import { WMTSGeoResource } from '../../../../../src/services/domain/geoResources'; 
 import { $injector } from '../../../../../src/injection';
