+++ test/modules/map/components/layerManager/LayerManager.test.js
@@ -1,10 +1,10 @@
 import { LayerManager } from '../../../../../src/modules/map/components/layerManager/LayerManager';
 import { Toggle } from '../../../../../src/modules/commons/components/toggle/Toggle';
-import { layersReducer, defaultLayerProperties } from '../../../../../src/modules/map/store/layers.reducer';
+import { layersReducer, defaultLayerProperties } from '../../../../../src/store/layers/layers.reducer';
 import { TestUtils } from '../../../../test-utils';
 import { $injector } from '../../../../../src/injection';
 import { LayerItem } from '../../../../../src/modules/map/components/layerManager/LayerItem';
-import { modifyLayer } from '../../../../../src/modules/map/store/layers.action';
+import { modifyLayer } from '../../../../../src/store/layers/layers.action';
 
 window.customElements.define(Toggle.tag, Toggle);
 window.customElements.define(LayerItem.tag, LayerItem);
