+++ test/modules/map/store/GeolocationPlugin.test.js
@@ -1,7 +1,7 @@
 import { GeolocationPlugin, GEOLOCATION_LAYER_ID } from '../../../../src/modules/map/store/GeolocationPlugin';
 import { activate, deactivate, setTracking } from '../../../../src/modules/map/store/geolocation.action';
 import { TestUtils } from '../../../test-utils.js';
-import { layersReducer } from '../../../../src/modules/map/store/layers.reducer';
+import { layersReducer } from '../../../../src/store/layers/layers.reducer';
 import { geolocationReducer } from '../../../../src/modules/map/store/geolocation.reducer';
 import { $injector } from '../../../../src/injection';
 import { positionReducer } from '../../../../src/store/position/position.reducer';
