+++ test/service/ShareService.test.js
@@ -1,6 +1,6 @@
 import { $injector } from '../../src/injection';
-import { addLayer } from '../../src/modules/map/store/layers.action';
-import { layersReducer } from '../../src/modules/map/store/layers.reducer';
+import { addLayer } from '../../src/store/layers/layers.action';
+import { layersReducer } from '../../src/store/layers/layers.reducer';
 import { changeZoomAndCenter } from '../../src/store/position/position.action';
 import { positionReducer } from '../../src/store/position/position.reducer';
 import { setCurrent } from '../../src/modules/topics/store/topics.action';
